chore(webpack): many optimizations
-compression in the build step... both brotli and then gzip as fallback

-flags manuall enabled and finetuned, specifically chunking, to split the bundle filesizes for more effective caching/update rollout

-terser clears all comments

-script loading default set to defer

-general preparation for dynamic imports and lazy loading

-pkg type set to module

-setting to module, in turn required webpack.config.js to be set to a cjs type... however to avoid needing to send the config flag to each webpack call, a symbolic link has been created in js' place pointing to the cjs
