+++ package.json
@@ -5,6 +5,7 @@
     "node": ">=14.0.0",
     "npm": ">=6.0.0 <7.0.0"
   },
+  "type": "module",
   "devDependencies": {
     "jest": "^26.6.3",
     "serve": "^11.0.0",
@@ -24,6 +25,7 @@
     "bs58check": "^2.1.2",
     "cbor": "^5.1.0",
     "clean-webpack-plugin": "^3.0.0",
+    "compression-webpack-plugin": "^6.1.1",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.1",
     "file-loader": "^6.2.0",
@@ -34,7 +36,7 @@
     "style-loader": "^0.23.1",
     "svelte": "^3.0.0",
     "svelte-loader": "^2.13.6",
-    "webpack": "^4.30.0",
+    "webpack": "^4.46.0",
     "webpack-cli": "^3.3.0",
     "workbox-webpack-plugin": "^6.1.2",
     "worker-loader": "^3.0.2"
